requirements.txt
matplotlib
onnxruntime
pveagle>=1.0.0
pyannote.audio
pyannote.metrics
requests
scipy
soundfile
speechbrain
torch
tqdm